修改library包名，处理包名不合法的问题
